feat(OmokGame): 게임을 진행한다
